Extract version command to common location
